+++ CHANGELOG.md
@@ -7,16 +7,27 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased](https://github.com/askap-vast/vast-post-processing/compare/v2.1.1...HEAD)
 
 ### Added
+- Added a filter that removes sources at the edges of observations when calculating the astromteric and flux corrections. These sources will be cropped out later. [#121](https://github.com/askap-vast/vast-post-processing/pull/121)
+- Added a method that uses a linear fit with a Huber regressor to calculate the flux correction factor. [#123](https://github.com/askap-vast/vast-post-processing/pull/123)
+- Added statsmodels to the poetry file.
 
 ### Changed
+- Changed the way the flux offset factor is calculated, defaults to Huber regressor.
 
 ### Fixed
-
 - Fixed error in extending source filtering that reversed numerator and denominator [#120](https://github.com/askap-vast/vast-post-processing/pull/120)
 
 ### Removed
+- In crossmatch.py, removed the calculate_flux_offsets() function. Replaced with calculate_flux_offsets_median() [#122](https://github.com/askap-vast/vast-post-processing/pull/122)
+- In crossmatch.py, removed the faulty median_abs_deviation() function. Replaced with astropy.stats.mad_std [#122](https://github.com/askap-vast/vast-post-processing/pull/122)
+- [#122](https://github.com/askap-vast/vast-post-processing/pull/122): removed:  Changed the way the flux offset factor is calculated. Instead of using the slope of the linear fit to the flux_int vs. flux_int_reference plane, we now use the median of the distribution flux_int/flux_int_reference. Redundant now that we use the Huber regressor. Pull request was not implemented, closed.
 
 ### List of PRs
+- [#123](https://github.com/askap-vast/vast-post-processing/pull/123): added: Added a method that uses a linear fit with a Huber regressor to calculate the flux correction factor. Made this the default.
+
+- [#122](https://github.com/askap-vast/vast-post-processing/pull/122): changed:  Changed the way the flux offset factor is calculated. Instead of using the slope of the linear fit to the flux_int vs. flux_int_reference plane, we now use the median of the distribution flux_int/flux_int_reference. This is similar to how the correction is calculate for the astrometry. Both flux and astrometry correction now use the median absolute deviation as an error. Pull request was not implemented, closed.
+
+- [#121](https://github.com/askap-vast/vast-post-processing/pull/121): added: Added a filter that removes sources at the edges of observations when calculating the astromteric and flux corrections. These sources will be cropped out later.
 
 - [#120](https://github.com/askap-vast/vast-post-processing/pull/120): fix: Fixed error in extending source filtering that reversed numerator and denominator